Fix missing LICENSE on PyPI
Add a symbolic link to the LICENSE file in the parent directory so that
it (already named in MANIFEST.in) appears in the PyPI source and wheel
distributions.

This is particularly important because the chosen MIT license requires
that the copyright statement and license text be reproduced in
derivative works; to comply with the license, library users need it to
be included, not just referenced.